refactor(core): remove unneeded LocalClient functionality, cleanup migrations
