+++ renku/core/commands/checks/datasets.py
@@ -21,7 +21,7 @@ from collections import defaultdict
 
 import click
 
-from renku.core.utils.migrate import get_pre_0_3_4_datasets_metadata
+from renku.core.management.migrations.utils import get_pre_0_3_4_datasets_metadata
 
 from ..echo import WARNING
