+++ tests/core/commands/test_client.py
@@ -56,25 +56,20 @@ def test_safe_class_attributes(tmpdir):
 
     # NOTE: attributes that are allowed on LocalClient
     safe_attributes = [
-        "ACTIVITY_INDEX",
         "CACHE",
         "CONFIG_NAME",
         "DATABASE_PATH",
-        "DATASETS",
         "DATASET_IMAGES",
-        "DATASETS_PROVENANCE",
         "DATA_DIR_CONFIG_KEY",
         "DEPENDENCY_GRAPH",
         "DOCKERFILE",
         "LOCK_SUFFIX",
         "METADATA",
         "POINTERS",
-        "PROVENANCE_GRAPH",
         "RENKU_LFS_IGNORE_PATH",
         "RENKU_PROTECTED_PATHS",
         "SUPPORTED_SCHEMES",
         "TEMPLATE_CHECKSUMS",
-        "WORKFLOW",
         "_CMD_STORAGE_CHECKOUT",
         "_CMD_STORAGE_CLEAN",
         "_CMD_STORAGE_INSTALL",
@@ -87,9 +82,7 @@ def test_safe_class_attributes(tmpdir):
         "_CMD_STORAGE_UNTRACK",
         "_LFS_HEADER",
         "_database",
-        "_datasets_provenance",
         "_global_config_dir",
-        "_temporary_datasets_path",
     ]
 
     client1 = LocalClient(str(tmpdir.mkdir("project1")))
